Moved functions out of iris_engine.access_control.utils into the business layer
