+++ source/app/business/access_controls.py
@@ -15,9 +15,101 @@
 #  You should have received a copy of the GNU Lesser General Public License
 #  along with this program; if not, write to the Free Software Foundation,
 #  Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+from sqlalchemy import and_
+
+from app import db
+from app.datamgmt.manage.manage_access_control_db import get_case_effective_access
+from app.datamgmt.manage.manage_access_control_db import check_ua_case_client
 from app.iris_engine.access_control.iris_user import iris_current_user
-from app.iris_engine.access_control.utils import ac_fast_check_user_has_case_access
+from app.logger import logger
+from app.models.authorization import UserCaseAccess
+from app.models.authorization import UserCaseEffectiveAccess
+from app.models.authorization import CaseAccessLevel
+from app.models.authorization import ac_flag_match_mask
 
 
 def ac_fast_check_current_user_has_case_access(cid, access_level):
     return ac_fast_check_user_has_case_access(iris_current_user.id, cid, access_level)
+
+
+def set_user_case_access(user_id, case_id, access_level):
+
+    uca = UserCaseAccess.query.filter(
+        UserCaseAccess.user_id == user_id,
+        UserCaseAccess.case_id == case_id
+    ).all()
+
+    if len(uca) > 1:
+        for u in uca:
+            db.session.delete(u)
+        db.session.commit()
+        uca = None
+
+    if not uca:
+        uca = UserCaseAccess()
+        uca.user_id = user_id
+        uca.case_id = case_id
+        uca.access_level = access_level
+        db.session.add(uca)
+    else:
+        uca[0].access_level = access_level
+
+    db.session.commit()
+
+    set_case_effective_access_for_user(user_id, case_id, access_level)
+
+
+def set_case_effective_access_for_user(user_id, case_id, access_level: int):
+    """
+    Set a case access from a user
+    """
+
+    uac = UserCaseEffectiveAccess.query.where(and_(
+        UserCaseEffectiveAccess.user_id == user_id,
+        UserCaseEffectiveAccess.case_id == case_id
+    )).all()
+
+    if len(uac) > 1:
+        logger.error(f'Multiple access found for user {user_id} and case {case_id}')
+        for u in uac:
+            db.session.delete(u)
+        db.session.commit()
+
+        uac = UserCaseEffectiveAccess()
+        uac.user_id = user_id
+        uac.case_id = case_id
+        uac.access_level = access_level
+        db.session.add(uac)
+
+    elif len(uac) == 1:
+        uac = uac[0]
+        uac.access_level = access_level
+
+    db.session.commit()
+
+
+def ac_fast_check_user_has_case_access(user_id, cid, expected_access_levels: list[CaseAccessLevel]):
+    """
+    Checks the user has access to the case with at least one of the access_level
+    if the user has access, returns the access level of the user to the case
+    Returns None otherwise
+    """
+    access_level = get_case_effective_access(user_id, cid)
+
+    if not access_level:
+        # The user has no direct access, check if he is part of the client
+        access_level = check_ua_case_client(user_id, cid)
+        if not access_level:
+            return None
+        set_case_effective_access_for_user(user_id, cid, access_level)
+
+        return access_level
+
+    if ac_flag_match_mask(access_level, CaseAccessLevel.deny_all.value):
+        return None
+
+    for acl in expected_access_levels:
+        if ac_flag_match_mask(access_level, acl.value):
+            return access_level
+
+    return None
